chore: extract check line
