+++ packages/blocks/src/page-block/utils/position.ts
@@ -7,7 +7,7 @@ import {
   resetNativeSelection,
   SelectionEvent,
 } from '../../__internal__/index.js';
-import { isAtLineEdge } from '../../__internal__/rich-text/rich-text-operations.js';
+import { isAtLineEdge } from '../../__internal__/utils/check-line.js';
 
 export function repairContextMenuRange(e: SelectionEvent) {
   const selection = window.getSelection() as Selection;
